+++ @planetarium/account-web3-secret-storage/src/KeyId.ts
@@ -1,3 +1,18 @@
+import { Crypto } from "@peculiar/webcrypto";
+
+const crypto = new Crypto();
+
 export type KeyId = string;
+
+const UUID_PATTERN =
+  /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i;
+
+export function isKeyId(keyId: string): keyId is KeyId {
+  return !!keyId.match(UUID_PATTERN);
+}
+
+export function generateKeyId(): KeyId {
+  return crypto.randomUUID();
+}
 
 export default KeyId;