+++ @planetarium/account-web3-secret-storage/src/index.ts
@@ -1,5 +1,10 @@
 export { KeyId } from "./KeyId.js";
 export { PassphraseEntry } from "./PassphraseEntry.js";
 export { TtyPassphraseEntry } from "./TtyPassphraseEntry.js";
+export {
+  type Web3AccountOptions as DecryptionOptions,
+  WeakPrivateKeyError,
+  Web3Account,
+  type Web3KeyObject,
+} from "./Web3Account.js";
 export { getDefaultWeb3KeyStorePath, Web3KeyStore } from "./Web3KeyStore.js";
-export { Web3Account, type Web3KeyObject } from "./Web3Account.js";