+++ CHANGES.md
@@ -131,6 +131,14 @@ To be released.
     commands related to *Libplanet.Explorer* project.  [[#2613]]
      -  Added `IndexCommand` Cocona command class.
  -  (Libplanet.Explorer) Added `states` query in `StateQuery`.  [[#3080]]
+ -  (@planetarium/account) Added `Address.equals()` method.  [[#3071]]
+ -  (@planetarium/account) Added `Address.isAddressOf()` method.  [[#3071]]
+ -  (@planetarium/account) Added `PublicKey.equals()` method.  [[#3071]]
+ -  (@planetarium/account-web3-secret-storage) Added `DecryptionOptions`
+    interface.  [[#3071]]
+     -  Added `Web3KeyStoreOptions.decryptionOptions` attribute.
+ -  (@planetarium/account-web3-secret-storage) Added `WeakPrivateKeyError`
+    class.  [[#3071]]
 
 ### Behavioral changes
 
@@ -198,11 +206,17 @@ To be released.
      -  `transactionResult` query in `TransactionQuery` now uses
         `IBlockChainIndex` if available.
 
+ -  (@planetarium/account-web3-secret-storage) `Web3KeyStore` now supports
+    Scrypt as one of KDF functions besides PBKDF2.  [[#3071]]
+
 ### Bug fixes
 
  -  In `PreVote` block validation, `Context<T>.IsValid()`, validate the block
     header and also the block content (e.g., Tx nonces, Policy defined
     validation rules, or state root hash.)  [[#2973], [#2996]]
+ -  (@planetarium/account-web3-secret-storage) Fixed a bug where
+    `Web3KeyStore.import()` and `Web3KeyStore.generate()` methods had written
+    JSON field `"Crypto"` instead of `"crypto"` in the key file.  [[#3071]]
 
 ### Dependencies
 
@@ -235,6 +249,7 @@ To be released.
 [#3061]: https://github.com/planetarium/libplanet/pull/3061
 [#3067]: https://github.com/planetarium/libplanet/pull/3067
 [#3069]: https://github.com/planetarium/libplanet/pull/3069
+[#3071]: https://github.com/planetarium/libplanet/pull/3071
 [#3072]: https://github.com/planetarium/libplanet/pull/3072
 [#3074]: https://github.com/planetarium/libplanet/pull/3074
 [#3077]: https://github.com/planetarium/libplanet/pull/3077